pyproject.toml
[project]
name = "anilist-link"
version = "0.1.0"
description = "Self-hosted bridge between AniList and media platforms (Crunchyroll, Plex, Jellyfin)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "apscheduler>=3.10.0,<4.0.0",
    "rapidfuzz>=3.0.0",
    "selenium>=4.15.0,<4.27.0",
    "undetected-chromedriver>=3.5.4",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"src/Web/Routes/Auth.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "apscheduler.*",
    "undetected_chromedriver.*",
]
ignore_missing_imports = true
